# Performance: Cache NonCircularContractResolver