@NickCraver 2022-01-19
@brettsam @mathewc Not sure on merge policies for the SDK - are we good to fold this in?